FIXED assignTargets!!! (kind of)
assignTargets now works, most of the time! I made a copy of the teams list so that it can be reused multiple times. More importantly, I realized that I needed to reset firstPerson to a random person from teamsCopy. Before, it didn't work because firstPersn was a person outside of teamsCopy. Wow it feels good when the code works*

*works for the most part, sometimes it gets stuck in an infinite loop, still got to figure that out!